+++ packages/lib-sourcify/src/Compilation/SolidityCompilation.ts
@@ -125,7 +125,7 @@ export class SolidityCompilation extends AbstractCompilation {
           this.auxdataStyle,
         );
 
-        if (runtimeAuxdataCbor) {
+        if (runtimeAuxdataCbor && runtimeCborLengthHex !== undefined) {
           const auxdataFromRawRuntimeBytecode = `${runtimeAuxdataCbor}${runtimeCborLengthHex}`;
           this._runtimeBytecodeCborAuxdata = {
             '1': {
@@ -148,7 +148,7 @@ export class SolidityCompilation extends AbstractCompilation {
           this.auxdataStyle,
         );
 
-        if (creationAuxdataCbor) {
+        if (creationAuxdataCbor && creationCborLengthHex !== undefined) {
           const auxdataFromRawCreationBytecode = `${creationAuxdataCbor}${creationCborLengthHex}`;
           this._creationBytecodeCborAuxdata = {
             '1': {
@@ -187,8 +187,10 @@ export class SolidityCompilation extends AbstractCompilation {
           this.auxdataStyle,
         );
 
-        if (!runtimeAuxdataCbor) {
-          throw new Error('runtimeAuxdataCbor is undefined');
+        if (!runtimeAuxdataCbor || runtimeCborLengthHex === undefined) {
+          throw new Error(
+            'runtimeAuxdataCbor or runtimeCborLengthHex is undefined',
+          );
         }
 
         const auxdataFromRawRuntimeBytecode = `${runtimeAuxdataCbor}${runtimeCborLengthHex}`;
@@ -211,7 +213,7 @@ export class SolidityCompilation extends AbstractCompilation {
         );
 
         // If we can find the auxdata at the end of the bytecode return; otherwise continue with `generateEditedContract`
-        if (creationAuxdataCbor) {
+        if (creationAuxdataCbor && creationCborLengthHex !== undefined) {
           const auxdataFromRawCreationBytecode = `${creationAuxdataCbor}${creationCborLengthHex}`;
           // we divide by 2 because we store the length in bytes (without 0x)
           this._creationBytecodeCborAuxdata = {